server/api/serializers/product_serializer.py
from rest_framework import serializers

from api.models.product_model import Product


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = "__all__"
